Allow nonspatial layer import (#159)
* Allow nonspatial layer import

This commit includes minor changes to the importer to allow layers
without geometry information. This includes two things: First, we relax
the assumption that a geometry column must be present and let the
actual handler implementation decide to add the dimension lco option.
Second, we introduce a prepare_import step to the interface. Here, any
customizations can be done before the actual import starts.

Actual implementation of the datapackage handler moved to the contribs
repository and can be installed as mere dependency.

* Tests if prepare_import is called